app/requirements.txt
# ipython
# pandas
# pyspark
requests
kafka-python